+++ crates/dash_rt_http/src/lib.rs
@@ -9,8 +9,8 @@ use dash_rt::event::EventMessage;
 use dash_rt::module::ModuleLoader;
 use dash_rt::state::State;
 use dash_vm::delegate;
-use dash_vm::gc2::persistent::Persistent;
-use dash_vm::gc2::trace::Trace;
+use dash_vm::gc::persistent::Persistent;
+use dash_vm::gc::trace::Trace;
 use dash_vm::local::LocalScope;
 use dash_vm::throw;
 use dash_vm::value::function::native::CallContext;
