+++ crates/dash_vm/src/gc/handle.rs
@@ -1,18 +1,18 @@
-use std::{
-    cell::{Cell, RefCell},
-    fmt::Debug,
-    hash::Hash,
-    ops::Deref,
-    ptr::NonNull,
-};
+use std::cell::Cell;
+use std::hash::Hash;
+use std::ops::Deref;
+use std::ptr::NonNull;
 
 use bitflags::bitflags;
+
+use crate::value::object::Object;
 
 use super::trace::Trace;
 
 bitflags! {
     #[derive(Default)]
     struct HandleFlagsInner: u8 {
+        /// Whether the node has been visited in the last mark phase or not.
         const MARKED_VISITED = 1 << 0;
         const VM_DETACHED = 1 << 1;
     }
@@ -50,22 +50,38 @@ impl HandleFlags {
     }
 }
 
-#[derive(Debug)]
-pub struct InnerHandle<T: ?Sized> {
+pub struct GcNode<T: ?Sized> {
     pub(crate) flags: HandleFlags,
     /// Persistent<T> reference count
     pub(crate) refcount: Cell<u64>,
-    pub(crate) value: Box<T>,
-}
-
-impl<T: ?Sized> InnerHandle<T> {
-    pub fn ref_count(&self) -> u64 {
-        self.refcount.get()
-    }
+    /// The next pointer in the linked list of nodes
+    pub(crate) next: Option<NonNull<GcNode<dyn Object>>>,
+    pub(crate) value: T,
 }
 
 #[derive(Debug)]
-pub struct Handle<T: ?Sized>(NonNull<InnerHandle<T>>);
+pub struct Handle<T: ?Sized>(NonNull<GcNode<T>>);
+
+impl<T: ?Sized> Handle<T> {
+    /// # Safety
+    /// The given [`NonNull`] pointer must point to a valid [`InnerHandle`]
+    pub unsafe fn from_raw(ptr: NonNull<GcNode<T>>) -> Self {
+        Self(ptr)
+    }
+
+    pub fn into_raw(ptr: Self) -> NonNull<GcNode<T>> {
+        ptr.0
+    }
+}
+impl<T: Object + 'static> Handle<T> {
+    pub fn into_dyn(self) -> Handle<dyn Object> {
+        let ptr = Handle::into_raw(self);
+        // SAFETY: `T: Object` bound makes this cast safe. once CoerceUnsized is stable, we can use that instead.
+        unsafe { Handle::<dyn Object>::from_raw(ptr) }
+    }
+}
+
+impl<T: ?Sized> Eq for Handle<T> {}
 
 impl<T: ?Sized> PartialEq for Handle<T> {
     fn eq(&self, other: &Self) -> bool {
@@ -73,28 +89,40 @@ impl<T: ?Sized> PartialEq for Handle<T> {
     }
 }
 
-impl<T: ?Sized> Eq for Handle<T> {}
-
 impl<T: ?Sized> Clone for Handle<T> {
     fn clone(&self) -> Self {
         Self(self.0)
     }
 }
 
-impl<T: ?Sized> Handle<T> {
-    /// # Safety
-    /// The given [`NonNull`] pointer must point to a valid [`InnerHandle`]
-    pub unsafe fn new(ptr: NonNull<InnerHandle<T>>) -> Self {
-        Handle(ptr)
+// FIXME: this is severly unsound and hard to fix: this Handle can be smuggled because it's not tied to any reference
+// and later, when its GC goes out of scope, it will be freed, even though it's still alive
+impl<T: ?Sized> Deref for Handle<T> {
+    type Target = T;
+    fn deref(&self) -> &T {
+        unsafe { &(*self.as_ptr()).value }
     }
+}
 
-    pub fn as_ptr(&self) -> *mut InnerHandle<T> {
+impl<T: ?Sized> Handle<T> {
+    pub fn as_ptr(&self) -> *mut GcNode<T> {
         self.0.as_ptr()
     }
+}
 
-    pub fn replace(&mut self, new: Box<T>) {
-        let t = unsafe { self.0.as_mut() };
-        t.value = new;
+impl Handle<dyn Object> {
+    pub fn cast_handle<U: 'static>(&self) -> Option<Handle<U>> {
+        if self.as_any().is::<U>() {
+            Some(Handle(self.0.cast()))
+        } else {
+            None
+        }
+    }
+}
+
+impl<T: ?Sized> Hash for Handle<T> {
+    fn hash<H: std::hash::Hasher>(&self, state: &mut H) {
+        self.as_ptr().hash(state);
     }
 }
 
@@ -110,27 +138,5 @@ unsafe impl<T: ?Sized + Trace> Trace for Handle<T> {
         };
 
         T::trace(self);
-    }
-}
-
-unsafe impl<T: ?Sized + Trace> Trace for RefCell<T> {
-    fn trace(&self) {
-        T::trace(&RefCell::borrow(self));
-    }
-}
-
-// FIXME: this is severly unsound and hard to fix: this Handle can be smuggled because it's not tied to any reference
-// and later, when its GC goes out of scope, it will be freed, even though it's still alive
-impl<T: ?Sized> Deref for Handle<T> {
-    type Target = T;
-
-    fn deref(&self) -> &Self::Target {
-        unsafe { &self.0.as_ref().value }
-    }
-}
-
-impl<T: ?Sized> Hash for Handle<T> {
-    fn hash<H: std::hash::Hasher>(&self, state: &mut H) {
-        self.as_ptr().hash(state);
     }
 }
