+++ crates/dash_vm/src/gc/trace.rs
@@ -1,4 +1,5 @@
 use std::cell::Cell;
+use std::cell::RefCell;
 use std::collections::HashSet;
 use std::rc::Rc;
 
@@ -13,7 +14,7 @@ use crate::value::typedarray::TypedArrayKind;
 
 /// # Safety
 /// Implementors of this trait must provide a valid trace implementation
-/// by calling any possible, reachable [`super::Handle`]
+/// by calling any possible, reachable [`super::Handle`]s
 pub unsafe trait Trace {
     fn trace(&self);
 }
@@ -57,6 +58,18 @@ unsafe impl<T: Trace + ?Sized> Trace for Rc<T> {
 unsafe impl<T: Trace + Copy> Trace for Cell<T> {
     fn trace(&self) {
         Cell::get(self).trace();
+    }
+}
+
+unsafe impl<T: Trace + ?Sized> Trace for Box<T> {
+    fn trace(&self) {
+        T::trace(self)
+    }
+}
+
+unsafe impl<T: ?Sized + Trace> Trace for RefCell<T> {
+    fn trace(&self) {
+        T::trace(&RefCell::borrow(self));
     }
 }
