+++ crates/dash_vm/src/statics.rs
@@ -1,7 +1,7 @@
 use dash_proc_macro::Trace;
 
-use crate::gc2::handle::Handle;
-use crate::gc2::Gc;
+use crate::gc::handle::Handle;
+use crate::gc::Gc;
 use crate::js_std;
 use crate::value::error::AggregateError;
 use crate::value::error::EvalError;
