+++ crates/dash_vm/src/value/array.rs
@@ -5,7 +5,7 @@ use std::cell::RefCell;
 use dash_proc_macro::Trace;
 
 use crate::delegate;
-use crate::gc2::handle::Handle;
+use crate::gc::handle::Handle;
 use crate::local::LocalScope;
 use crate::throw;
 use crate::Vm;
