+++ crates/dash_vm/src/value/arraybuffer.rs
@@ -4,7 +4,7 @@ use std::cell::Cell;
 use dash_proc_macro::Trace;
 
 use crate::delegate;
-use crate::gc2::handle::Handle;
+use crate::gc::handle::Handle;
 use crate::local::LocalScope;
 use crate::Vm;
