+++ crates/dash_vm/src/value/error.rs
@@ -5,7 +5,7 @@ use std::rc::Rc;
 use dash_proc_macro::Trace;
 
 use crate::delegate;
-use crate::gc2::handle::Handle;
+use crate::gc::handle::Handle;
 use crate::local::LocalScope;
 use crate::Vm;
