+++ crates/dash_vm/src/value/function/async.rs
@@ -1,7 +1,7 @@
 use dash_proc_macro::Trace;
 
 use crate::delegate;
-use crate::gc2::handle::Handle;
+use crate::gc::handle::Handle;
 use crate::local::LocalScope;
 use crate::value::object::NamedObject;
 use crate::value::object::Object;
