+++ crates/dash_vm/src/value/function/user.rs
@@ -5,7 +5,7 @@ use dash_proc_macro::Trace;
 
 use crate::dispatch::HandleResult;
 use crate::frame::Frame;
-use crate::gc2::handle::Handle;
+use crate::gc::handle::Handle;
 use crate::local::LocalScope;
 use crate::value::ExternalValue;
 use crate::value::Value;
