+++ crates/dash_vm/src/value/ops/abstractions/conversions.rs
@@ -1,6 +1,6 @@
 use std::rc::Rc;
 
-use crate::gc2::handle::Handle;
+use crate::gc::handle::Handle;
 use crate::local::LocalScope;
 use crate::throw;
 use crate::value::boxed::Boolean;
