+++ crates/dash_vm/src/value/promise.rs
@@ -3,8 +3,8 @@ use std::cell::RefCell;
 
 use dash_proc_macro::Trace;
 
-use crate::gc2::handle::Handle;
-use crate::gc2::trace::Trace;
+use crate::gc::handle::Handle;
+use crate::gc::trace::Trace;
 use crate::local::LocalScope;
 use crate::PromiseAction;
 use crate::Vm;
