+++ crates/dash_vm/src/value/typedarray.rs
@@ -2,7 +2,7 @@ use std::any::Any;
 
 use dash_proc_macro::Trace;
 
-use crate::gc2::handle::Handle;
+use crate::gc::handle::Handle;
 use crate::local::LocalScope;
 use crate::Vm;
